2 and 3d code split more strcitly due to new version of eigen3
