--- src/for_3D_build/geometries/geometric_shape_3d.cpp
@@ -1,15 +1,0 @@
-#include "geometric_shape.h"
-
-#include "io_environment.h"
-#include "triangle_mesh_shape.h"
-
-namespace SPH
-{
-//=================================================================================================//
-void GeometricShapeBox::writeGeometricShapeBoxToVtp(Real scale_factor)
-{
-    TriangleMeshShapeBrick shape(HalfSize(), 1, Vecd::Zero(), Name());
-    shape.writTriangleMeshShapeToVtp(getTransform(), scale_factor);
-}
-//=================================================================================================//
-} // namespace SPH
