+++ src/shared/common/vector_functions.h
@@ -31,8 +31,10 @@
 #include "data_type.h"
 
 #include <Eigen/Geometry>
-#include <Eigen/SVD>
 #include <Eigen/LU>
+#include <Eigen/SVD>
+
+#include <cassert>
 
 namespace SPH
 {
@@ -154,7 +156,6 @@ template <int Dim1, int Dim2, typename ComponentFunction, typename... Args>
 void for_each_component(const Eigen::Matrix<Real, Dim1, Dim2> &input,
                         const ComponentFunction &function, Args &&...args)
 {
-    Eigen::Matrix<Real, Dim1, Dim2> output;
     for (int i = 0; i < Dim1; ++i)
         for (int j = 0; j < Dim2; ++j)
             function(input(i, j), std::forward<Args>(args)(i, j)...);
