+++ src/shared/geometries/geometric_element.h
@@ -75,26 +75,5 @@ class GeometricBall
   protected:
     Real radius_;
 };
-
-class GeometricCylinder
-{
-  public:
-    explicit GeometricCylinder(Real radius, Real halflength);
-    ~GeometricCylinder() {};
-
-    bool checkContain(const Vecd &probe_point)
-    {
-        if (ABS(probe_point[0]) > halflength_)
-            return false;
-        return probe_point.tail(Dimensions - 1).norm() <= radius_;
-    };
-
-    Vecd findClosestPoint(const Vecd &probe_point);
-    BoundingBoxd findBounds();
-
-  protected:
-    Real radius_;
-    Real halflength_;
-};
 } // namespace SPH
 #endif // GEOMETRIC_ELEMENT_H
