+++ src/shared/geometries/geometric_shape.h
@@ -35,6 +35,8 @@
 
 namespace SPH
 {
+using TransformGeometryBox = TransformGeometry<GeometricBox>;
+
 class GeometricShapeBox : public TransformShape<GeometricBox>
 {
   public:
@@ -44,7 +46,7 @@ class GeometricShapeBox : public TransformShape<GeometricBox>
                       const std::string &name = "GeometricShapeBox");
     explicit GeometricShapeBox(const BoundingBoxd &bounding_box,
                                const std::string &name = "GeometricShapeBox");
-    virtual ~GeometricShapeBox() {};
+    virtual ~GeometricShapeBox(){};
     TransformGeometryBox getExpandedBox(Real expansion) const;
     void writeGeometricShapeBoxToVtp(Real scale_factor = 1.0);
 };
@@ -56,21 +58,13 @@ class GeometricShapeBall : public GeometricBall, public Shape
   public:
     explicit GeometricShapeBall(const Vecd &center, Real radius,
                                 const std::string &name = "GeometricShapeBall");
-    virtual ~GeometricShapeBall() {};
+    virtual ~GeometricShapeBall(){};
 
     virtual bool checkContain(const Vecd &probe_point, bool BOUNDARY_INCLUDED = true) override;
     virtual Vecd findClosestPoint(const Vecd &probe_point) override;
 
   protected:
     virtual BoundingBoxd findBounds() override;
-};
-
-class GeometricShapeCylinder : public TransformShape<GeometricCylinder>
-{
-  public:
-    GeometricShapeCylinder(const Transform &transform, Real radius, Real halflength,
-                           const std::string &name = "GeometricShapeCylinder");
-    virtual ~GeometricShapeCylinder() {};
 };
 } // namespace SPH
