--- tests/unit_tests_src/for_2D_build/geometries/test_2d_geometric_cylinder/CMakeLists.txt
@@ -1,16 +1,0 @@
-STRING( REGEX REPLACE ".*/(.*)" "\\1" CURRENT_FOLDER ${CMAKE_CURRENT_SOURCE_DIR} )
-PROJECT("${CURRENT_FOLDER}")
-
-SET(LIBRARY_OUTPUT_PATH ${PROJECT_BINARY_DIR}/lib)
-SET(EXECUTABLE_OUTPUT_PATH "${PROJECT_BINARY_DIR}/bin/")
-SET(BUILD_INPUT_PATH "${EXECUTABLE_OUTPUT_PATH}/input")
-SET(BUILD_RELOAD_PATH "${EXECUTABLE_OUTPUT_PATH}/reload")
-
-aux_source_directory(. DIR_SRCS)
-ADD_EXECUTABLE(${PROJECT_NAME} ${EXECUTABLE_OUTPUT_PATH} ${DIR_SRCS})
-target_link_libraries(${PROJECT_NAME} sphinxsys_2d GTest::gtest GTest::gtest_main)				 
-set_target_properties(${PROJECT_NAME} PROPERTIES VS_DEBUGGER_WORKING_DIRECTORY "${EXECUTABLE_OUTPUT_PATH}")
-
-add_test(NAME ${PROJECT_NAME} 
-		 COMMAND ${PROJECT_NAME}
-		 WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
