+++ tests/unit_tests_src/for_3D_build/geometrices/test_3d_geometric_cylinder/test_geometric_cylinder.cpp
@@ -1,5 +1,5 @@
-#include "geometric_shape.h"
-#include "transform_geometry.h"
+#include "geometric_element_and_shape_3d.h"
+
 #include <gtest/gtest.h>
 
 using namespace SPH;
